re-organize project structure of client app
